refactor: move utils up for re-use
