+++ src/data/utils/index.ts
@@ -1,3 +1,5 @@
 export * from "./categorize";
+export * from "./getRRULE";
+export * from "./getStartEndDates";
 export * from "./passwd";
 export * from "./readJsonAsync";
